// frontend/src/codegen/model/passkeySummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Flathub API
 * OpenAPI spec version: 0.1.0
 */
import type { PasskeySummaryLastUsedAt } from "./passkeySummaryLastUsedAt"

export interface PasskeySummary {
  id: number
  name: string
  created_at: string
  last_used_at: PasskeySummaryLastUsedAt
}